requirements.txt
# --- Core ---
Flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.31.0

# --- Speech-to-Text ---
faster-whisper==0.9.0
torch==2.2.0 --extra-index-url https://download.pytorch.org/whl/cpu
onnxruntime==1.17.1
soundfile==0.12.1
numpy==1.26.4

# --- Text-to-Speech ---
elevenlabs==0.2.27
gTTS==2.3.2
pydub==0.25.1

# --- AI & NLP ---
openai==0.27.8
nltk==3.8.1
langdetect==1.0.9

# --- Google Cloud ---
google-cloud-speech==2.21.0
google-cloud-translate==2.0.1

# --- Vapi ---
vapi-python==0.1.7